Update Karma configurations
